\name{sultan}
\docType{data}
\alias{sultan}
\title{RNA-seq data from humans in Sultan et al. (2008)}
\description{
This dataset represents RNA-seq data from humans in two conditions (Ramos B cell line and HEK293T), with two biological replicates per condition.
The ExpressionSet was downloaded from the ReCount online resource.
}
\usage{sultan}
\format{
An ExpressionSet named \code{sultan.eset} containing the phenotype data and 
expression data for the Sultan et al. (2008) experiment. Phenotype data may be
accessed using the \code{pData} function, and expression data may be accessed
using the \code{exprs} function.
}
\source{
ReCount online resource (http://bowtie-bio.sourceforge.net/recount).
}
\references{
A. C. Frazee, B. Langmead, and J. T. Leek. ReCount: a multi-experiment resource
of analysis-ready RNA-seq gene count datasets. BMC Bioinformatics, 12(449), 2011.

M. Sultan, M. H. Schulz, H. Richard, A. Magen, A. Klingenhoff, M. Scherf, M. Seifert,
T. Borodina, A. Soldatov, D. Parkhomchuk, D. Schmidt, S. O'Keefe, S. Haas,
M. Vingron, H. Lehrach, and M. L. Yaspo. A global view of gene activity and alternative 
splicing by deep sequencing of the human transcriptome. Science, 15(5891):956-60, 2008.
}
\keyword{datasets}

